requirements.txt
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
dbus-python==1.2.16
distro==1.4.0
einops==0.7.0
h5py==3.6.0
idna==3.3
numpy==1.22.3
Pillow==9.1.0
protobuf==3.20.1
PyGObject==3.36.0
PyYAML==6.0
requests==2.27.1
scipy==1.10.1
six==1.16.0
ssh-import-id==5.10
tensorboardX==2.5
torch==1.11.0+cu113
torchvision==0.12.0+cu113
typing_extensions==4.2.0
urllib3==1.26.9
